feat: Plugin API tweaks
